mnist_project/requirements.txt
click==8.0.3
matplotlib==3.5.1
numpy==1.22.0
python-dotenv==0.19.2
seaborn==0.11.2
setuptools==60.2.0
torch==1.10.1
tqdm==4.62.3
